Removed myNdla sharedpage. Refactored SharedFolder for new design
