+++ src/routes.tsx
@@ -74,9 +74,6 @@ export const routes = [
   "subject:subjectId/topic:topic1/topic:topic2/topic:topic3/topic:topic4/topic:topicId/resource:resourceId",
   "subject:subjectId/topic:topic1/topic:topic2/topic:topic3/topic:topic4/topic:topicId/resource:resourceId/:stepId",
   "minndla",
-  "minndla/folders/shared/:folderId",
-  "minndla/folders/shared/:subFolderId/:resourceId/",
-  "minndla/folders/shared/:folderId/:subFolderId/",
   ...privateRoutes,
 ];
